threadweaver-backend/requirements.txt
fastapi
uvicorn
pydantic
pyjwt

supabase
asyncpg

langchain>=0.4.0
langgraph
langchain-anthropic
langchain-text-splitters
langchain-openai

openai
anthropic

python-dotenv
httpx

pydantic-settings

pytest
pytest-mock
pytest-asyncio

python-multipart

"mcp[cli]"